elf.h : fix 32 bits elf

+++ include/elf.h
@@ -6,7 +6,7 @@
 //32-bit ELF base types
 typedef uint32_t Elf32_Addr;
 typedef uint16_t Elf32_Half;
-typedef uint64_t Elf32_Off;
+typedef uint32_t Elf32_Off;
 typedef int32_t  Elf32_Sword;
 typedef uint32_t Elf32_Word;
